(3/19/25) Commit #3
Added two classes that are members of OsScanner:

	User: Represents a User of the system with six attributes
	      - username : str
	      - hasPaswword : bool
              - userid : str
	      - groupid : str
	      - home_dir : str
	      - shell : str

	Group: Represents a group of the system with four attributes
	      - groupname : str
              - hasPassword : bool
              - group_id : str
              - members : list of strings

Added python virtual environment for installing pip packages that are listed in the requirements.txt

Changed the get_OsUsers function that is part of the OsScanner class. The get_OsUsers function is now responsible for populating
the OsUsers member list with User objects.

Changed the get_OsGroups function that is part of the OsScanner class. The get_OsGroups function will now populate the OsGroups member
with Group objects.

Created a Tkinter widget with the data returned from the get_OsUsers function. The root widget contains a notebook widget for adding tabs
with the first tab being a Tree widget. The OsUsers Tree widget has six columns represented by the member variables of the User class.

Need to put all the GUI statments in a class to clean up the code